requirements.txt
argon2-cffi==19.1.0
bcrypt==3.1.7
cffi==1.12.3
Django==2.2.3
djangorestframework==3.10.3
Faker==2.0.1
Pillow==6.1.0
psycopg2-binary==2.8.3
pycparser==2.19
Pygments==2.4.2
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
